[FIX] Redis deserialize
OPEN API로부터 얻은 데이터의 유효성을 검증하기 위해 Redis로부터 유효한 데이터인지 조회하는 로직에서 기존 redis template 의 value 값이 object고 jackson으로 직렬화, 역직렬화 구성했기 때문에 기존 데이터를 조회하지 못하는 버그가 발생.
그래서 추가적으로 ReactiveStringRedisTemplate을 빈으로 등록 후 해당 template을 통해 역직렬화해서 데이터 조회에 대한 버그를 수정함.

+++ src/main/java/com/lolsearcher/reactive/utils/RiotGamesDataCacheKeyUtils.java
@@ -13,7 +13,7 @@ public class RiotGamesDataCacheKeyUtils {
         return CHAMPION_KEY_PREFIX + championId;
     }
 
-    public static String getItemKey(long item) {
+    public static String getItemKey(int item) {
 
         return ITEM_KEY_PREFIX + item;
     }
@@ -23,7 +23,7 @@ public class RiotGamesDataCacheKeyUtils {
         return QUEUE_KEY_PREFIX + queueId;
     }
 
-    public static String getSpellKey(long spell) {
+    public static String getSpellKey(int spell) {
 
         return SPELL_KEY_PREFIX + spell;
     }
